Stop passing procdescs with payloads in interprocedural analysis
Summary:
This diff changes the return type of the `analyze_dependency` callback in
`InterproceduralAnalysis.t` from `(Procdesc.t * 'payload) option` to just `'payload option`.  The
procdesc was previously just passed for convenience, but can be removed now (as we are removing
procdescs from summaries).  This also involves changes to all of the uses of that callback
throughout checkers and analyzers.

Reviewed By: skcho

Differential Revision: D43313294

fbshipit-source-id: d24ccfefd3f4219e7fd8f047a55d4520139cc10a