[unnecessary copy] Move type of the source from issue description to the trace
Summary:
In general, when we report issues, we try to report only one issue per line and per quailfier (by ignoring the procnames). However, for some issue types like Unnecessary copies or Deadstore, the qualifer contains the type information so for templated functions, we were reporting on the same line for each type instantiation.

This diff removes the types from the qualifiers so that we only report one issue. The type is still useful so we move it to the trace instead.

This also makes the error message nicer as we don't have to have the huge types sometimes.

A similar thing can be done for deadstore (next diff).

Reviewed By: skcho

Differential Revision: D43771201

fbshipit-source-id: 3636db46b749e4295bcd8cfcf0bbbcfd59b7710f