Change the location of lineage edges that go into procedure formal Return
Summary:
Interpreting instructions within the Lineage analysis generates some edges. We attach locations to
those edges to be able to track back the instructions that generated them in the source. In most
cases, the location of an edge will match the location of the last write of the local labelling its
destination node, since edges are usually introduced to record the write of that variable. Call edges
are a notable exception; which remains the same in this diff.

The case of edges introduced to write into the special `$ret` variable that denotes the formal return
of a procedure were at odds with that description: they denote the writing of a (special) variable,
yet they were attached the location of the last write of their source nodes.

We change these edges by attaching them the location of the exit node of the procedure instead (which
happens to map to the definition line of that procedure in source code). In some sense, we hereby
consider that the exit node is the node where the return of te function is written.

We also slightly change the graph simplification to keep those new edges: indeed, on the example
function `f() -> X.`, we generate a `X -(1)-> return -(2)-> $ret` structure in the graph that used to
get simplified into `X -(1)-> $ret`, since we happened to keep the first edge when both of them were
standard `Copy` ones (both `(1)` and `(2)` were mapped to the same location anyway, because of the
aforementioned behaviour). We now keep the second edge if both of them are `Copy`, ensuring that the
edges going into the `$ret` special node are always located to the exit node, as with the general
invariant that Copy edges are always located to the last write of their destination.

Note that this means that edges going from special argument node of structured arguments into the
corresponding variables, for instance `[$arg0] -(Project#1)-> [arg0#1]`, will be kept during
simplification, whereas those of unstructured arguments, eg. `[$arg0] -(Copy)-> [arg0]`, may be
simplified away and replaced with a "following" Copy edge. That means some edges going from `$argN`
special nodes will be located in the entry of a procedure and some will not -- we could change that
in the future if that turns out to be an issue.

Reviewed By: rgrig

Differential Revision: D43745045

fbshipit-source-id: 0c78b9ce9d99be69c03b383f539ebc4a1a382f94