[clang] In the mockpointer substitution preanalysis remove the macro location from the new call
Summary:
We still have a link to the macro in the trace which is one of the things we want to avoid with this preanalysis. This diff fixes it.

We also needed to move the macro expanded preanalysis before the dynamic dispatch one because that was modifying the mockpointer calls, so we were still getting the link to the macro in the trace.

Reviewed By: ngorogiannis

Differential Revision: D43911575

fbshipit-source-id: 85e64dca9ac80bcb39f38837e398e5bdf64ede28