[RFC][unnecessary copy] Report on intermediates from non-pvar
Summary:
We didn't report on intermediates that are not pvars (ie. the ones returned from function calls). This diff reports on them.

An example test is added for making an intermediate copy via syntactically calling `std::move` on a class that doesn't have a move operator defined, hence clang calls the copy ctor. (D44056113 is going to improve error messages to provide more context rather than saying something is copied via std::move which might be puzzling).

Another example is `std::make_unique` function that makes a copy if passed a non-rvalue ref. Now, we can detect this in the header file. Ideally, we want to report this issue at the callsite of make-unique but that can be done later.

Reviewed By: skcho

Differential Revision: D43770036

fbshipit-source-id: 60b7879e6252e5e1268506045308215163d1f907