[topl] Introduce Topl latent issues.
Summary:
A Topl issue is latent iff it depends on some constraints being
satisfied. There are two types of constraints: in the Pulse state (the
'conditions') and in the Topl state (the 'pruned' constraint).

Previously, there was no distinction between latent and manifest issues,
and all were reported as TOPL_ERROR. Now, latent issues are reported as
TOPL_ERROR_LATENT, and manifest issues are reported as TOPL_ERROR.

This commit gets rid of a false positive (in `erlang/topl/fields`) but
also introduces many false negatives. That is because it exposes several
limitations/bugs in the existing simplifications for Topl constraints,
which will be addressed in subsequent diffs.

Reviewed By: hajduakos

Differential Revision: D43192572

fbshipit-source-id: 45808ba8ba26514c9de36db75571fb883e21da33