[misc] Add a pretty printer for Procdesc that prints nodes with instructions
Summary:
Pre-existing Procdesc pretty printer doesn't print instructions. This diff adds a pretty printer
that (weekly) topo-sorts nodes of the CFG and then pretty-prints each node with its instrs. The
result should be a reasonable linear representation of the Procdesc.

I plan to use this pretty-printer for testing in future diffs.

Reviewed By: jvillard

Differential Revision: D44031984

fbshipit-source-id: 00c4089bc3043b76819fc5c490125977919b452a