[clang] Adding the info to the procedure attributes of which method a block has been passed to
Summary:
This is just a refactoring, because we already had some information in the proc attributes, namely whether the block was passed to a method in a position with the NS_NO_ESCAPE attribute. So it was just a matter of generalising this to storing the method that the block is passed to in addition to the given flag.

This will be useful in the taint analysis. We want to be able to express the following as a source:

"the first argument of a block passed to method X"

which is not possible right now.

Reviewed By: ngorogiannis

Differential Revision: D44094748

fbshipit-source-id: fc2e7484d2c29115687abee9e9382830d2f00c32