[pulse] model for load of std::is_same_v
Summary:
Adding a new capability to pulse in order to model the value of particular
globals. The mechanism is more general and can be used to write models for
`Load` instructions in general. We need this to model `std::is_same_v`, whose
value can be deduced statically.

  https://en.cppreference.com/w/cpp/types/is_same

Alternatively we could probably have modelled the *initializer* that the
frontend generates for that global and relied on the mechanism to run these
initializers on load but the approach in this diff seems more direct (eg
D43838894 also moved away from that approach).

Reviewed By: ezgicicek

Differential Revision: D44030597

fbshipit-source-id: 87ef80ec47acbe2b4864a3bd840fc6f31c919d1c