[clang] very basic coroutines support
Summary:
Translate coroutine-related AST nodes. This is enough for the taint analysis to
follow certain examples but it doesn't try to accurately match the semantics of
coroutines (eg perform the right calls to the awaitable and promise, etc.). It
doesn't try either to erase coroutines completely and restory a sequential
semantics, this would be harder to do, especially for coroutines with multiple
"return values", eg `co_yield 3; co_yield 4; co_yield 5;`.

Changes:
- ASTExporter.h: add translations for specific nodes:
  - `CoroutineBodyStmt`: this replaces the body of functions that are coroutines
  - `Coreturn` for `co_return` instructions
  - `CoroutineSuspendExpr` for `co_await` and `co_yield` instructions
- add two builtins `__builtin_cxx_co_return` and `__builtin_cxx_co_await` to be
  called on `co_return` and `co_await`/`co_yield`, respectively
- cTrans.ml now translates the corresponding AST nodes so that we see the
  coroutine code

Reviewed By: ezgicicek

Differential Revision: D44063784

fbshipit-source-id: 19aa67ec6ef921e4db51a97f0185f739457309e2