[pulse] Added --pulse-force-continue
Summary:
If a function f is complicated, it might not have any summaries of the
type ContinueProgram. In that case, code that comes after a call to
`f()` is not analyzed. With the `--pulse-force-continue`, calls to `f()`
are treated *also* as calls to an unknown function. This may introduce
false positives, but more code gets analyzed and true positives may also
be uncovered. ("also" means, for example, that if `f()` had
LatentAbortProgram or ExceptionRaised summaries, they continue to be
used by the analysis as they would without the --pulse-force-continue
option. So, the option only adds behaviors, for those situations when
summaries of type ContinueProgram are missing)

Reviewed By: ezgicicek

Differential Revision: D43734567

fbshipit-source-id: a3156376e18da321e1a4c2a06545f0b3391e6495