[doli] 3rd step in making doli practical
Summary:
The next step in making doli practical.

In this step we make it possible that the analysis of `sil` files can see the summaries of `doli`-defined functions.

**STATUS**

1. :-) Several `doli` and `sil` files can be captured together

2.  :-) Matching a call with a doli-rule is successful when the identifier at the call and one function identifier in the rule are identical.

3.  :-) creating formal parameter types and formal parameter identifiers out of `doli` rules.

4. :-) :-) The summary of  the `doli`-defined function **is** expanded at the place of the call. :-) :-)

5. :-( But
    a) Capture of doli rules has no knowledge of types, or other functions defined in  `*sil` files, nor other doli rules in the same file.
   b) I cannot see the summaries created for the doli files.
   c) We have not "closed the cycle with Java", ie pulse should be able to analyse Java methods which call `doli`-defined funtions

**EVIDENCE FOR ABOVE**

Look at the `README` file in the `tests\sil\silAndDoli` directory.

facebook

**QUESTIONS**

**Q1.** I suppose that we rely on the HLL (Hight Level Language) compiler. Then,  `DoliToTextual.match` will match the types of the parameters exactly, and apply subtypes for the receiver type. This means that we will need to *provide the signatures of the library functions* to the HLL  compiler.

**THINGS TO ADDRESS NEXT**

A. **Close the cycle with Java** -- as in 5.c above.

B. Improve the matching of function calls with `doli`-defined functions, so that the **types of parameters are taken into account**. It seems to me that here we will require the corresponding types to be equal.

C: Check that `doli` rules which have **several [extended] signatures**, are **matched correctly**.

D. Capture of `doli` rules should have  **knowledge of `textual`-defined functions**. Possible approaches:

- D1: Extend the syntax of Doli so that we can define externally-defined (ie defined in`sil` functions)

- D2: change how `TextualVerification` does type-checking for function calls?

We should probably go for D1.

E. Capture of doli rules should have  **knowledge of type definitions**. Possible approaches:

- E1: type definitions should be part of `doli` syntax. However, then the `doli` programmer would need to learn how to map HLL types to `Textual` types.

- E2: type definitions should be read out of `Textual` files. Seems OK, but still, ultimately requires `doli` programmer to learn how to map HLL types to `Textual` types.

- E3: type definitions obtained out of the HLL definitions.

E3 seems to be where we want to get to, and the question is whether for this we need/want to go through E2.

F. **Consistency checks** of `doli`-rules. The number and names of parameters in all the signatures in one rule should be the same.  Also, if the signatures allow for different parameter types, then we probably need to capture the bodies with these different types separately, ie capture once per signature. **QUESTION** In the future, we may want to allow for several different summaries per `doli` rule,  to express the dependency on types of receiver and arguments.

**IMPROVEMENTS**

Here are things that we can improve, but am not sure whether they would improve the functionality of our system.

I1: Pass `Location.t` from `CombinedMenhir` to `Doli` (as currently for `textual`), and drop `Unknown` entries. This might help with 5.b, above.

I2:  Improve the creation of `Procdesc` in `DoliToTextual`,and in particular

- a. Make the `nodes` information more precise, and add the labels of these nodes, esp the last one -- not clear what its value should be, though.

 - b. Question: Should indeed the `qualified_name` in the created `Procdecl` be the name of the rule? Does it matter? Also, the  `qualified_name`s seems to include parameter types and receiver class.

I3. Programming Style:

- a: The `Makefile` has a lot of repetition. We should have a loop instead. SD tried to write such a loop, but got muddled with the variable substitutions in the loop and gave up.

- b: The creation of the `ProcDesc` should probably take place in `CombinedMenhir` rather than in `DoliToTextual`.

- c: Find and remove redundancy in the data structures (too difficult for SD, but may bring efficiency. And it goes beyond `doli`)

**THINGS FOR FUTURE**

**F1.** Generics? Packages? Paths in the names of types?

**F2.** Exceptions?

**F3.** Capture of `doli` rules should have  **knowledge of `doli`-defined functions**.  This is more demanding, because we would require the `capture`-phase to become aware of the matching semantics. We need more examples, in order to decide whether we need it.

Reviewed By: davidpichardie

Differential Revision: D42907071

fbshipit-source-id: 2d1b55c2ed54d97174b32f4987c8ea31f4af0bf0