[IR] Fix an infinite loop by improving procname's unique_id function
Summary: We had an infinite loop in ondemand analysis in the example in the test. The reason was an inconsistency to define what the procname's unique id properties were. For storing and loading in the DB, `Procname.pp_unique_id` was used, which didn't distinguish between an Objective-C class `A` and and Objective-C protocol `A`. However, a set of Procnames would consider these two procnames to be different, because they include different `Typ.Name` in them. So making the function `pp_unique_id` take into account not only the name of the class (or protocol) but it's type as well solved the problem.

Reviewed By: jvillard

Differential Revision:
D44303366

Privacy Context Container: L1122176

fbshipit-source-id: d7cf1740e1bdf8191c058dfd75c48ff5d5b73789