[pulse] display proc names of pvars from other procedures
Summary:
Sometimes variables belong to a procedure that's not the current proc and
currently there is no way to tell when debugging pulse abstract states. This
can cause very mysterious bugs, resulting from errors in a frontend, or from
capturing variables. Display the proc name that variables belong to when it's
not the expected one (i.e. the one currently under analysis).

Reviewed By: ngorogiannis

Differential Revision: D44299220

fbshipit-source-id: 4abafe86c9e861cd36b7b1b78f2643a31d3779f0