[misc] Limit task bar update to 25Hz
Summary:
TaskBar implementation doesn't limit the frequency of redrawing into console. Current setup in
ProcessPool makes the effective frequency be 'at least 12Hz but potentially on every update from a
worker'. When there are many workers and many updates, the progress bar can push a lot of traffic
into the terminal. When the terminal is multiplexed (e.g. tmux in control mode + iTerm 2), a busy
progress bar can affect responsiveness of other terminal windows.

This diff adds throttling to TaskBar.refresh so that the frequency is not more than 25Hz which
feels snappy enough (to contrast, 10Hz feels a bit sluggish).

Reviewed By: jvillard

Differential Revision: D44582162

fbshipit-source-id: 7c4d488602fd42ad98cdf515d6fa153152eb2949