[pulse] Not generate specialized proc with empty function param
Summary:
`ClosureSpecialization` specializes procedures when it receives a function arameter, i.e. constructing a procname, `WithFunctionParameters of t * FunctionParameters.t list`.

OTOH, it cannot know the concrete function arguments always:

(it is checked/filtered at https://github.com/facebook/infer/blob/8e114728af78c15b15a3bab3d36991aafc102eff/infer/src/absint/ClosureSpecialization.ml#L24-L28)

In such cases, an empty function parameter list can be given for constructing a specialized procedure.

This is not bad itself, but in another place, the specialized procedure with the empty function paramter list is addressed as if a non-specialized procedure, e.g. in `pp_unique_id`,

https://github.com/facebook/infer/blob/8e114728af78c15b15a3bab3d36991aafc102eff/infer/src/IR/Procname.ml#L1191-L1192

This introduced a problem when calling the specialized procedure with the empty function parameter list:

* It tries to find a summary of the callee.
* `pp_unique_id` (the function above) is used, so it returns a summary for non-specialized procedure.
* In the summary, the local variables, including formals, have non-specialized procedure name, which is not what Pulse's call semantics expects.
* It fails to match formals and actuals correctly.

To avoid the issue, this diff forces that a specialized procedure is constructed only when the function parameter list is not empty.

Reviewed By: ngorogiannis

Differential Revision: D44635767

fbshipit-source-id: 878a58a681b197608d1e775f54d493256d1adba6