[ondemand][specialization] Alias specialization by Summary specialization
Summary:
Pulse relies currently on specialization for alias arguments and closures. The approach is based on cloning a function and specialized it by code transformation. We propose an other approach that does not modify the function body, but instead strengthen the summary of the specialized function with new pre/post. The specialization is now encoded in the summary precondition.

We believe this approach will we more efficient, more elegant and then can be more easily used for devirtualization.

Reviewed By: jvillard

Differential Revision: D43981938

fbshipit-source-id: 6db220f393d1639cda94bc96ef1173ed4d1a64cc